POC - do not merge!
Session window update hysteresis experiment:

Instead of sending FLOW performatives when the LWM is crossed, use
(hack) the LWM as a "maximum delta" to rate limit FLOW frame
generation.

When the difference between what the remote thinks is the window
capacity and the actual (local) capacity is >= LWM then send a FLOW
frame.

+++ c/src/core/buffer.c
@@ -35,9 +35,12 @@
 
 struct pn_buffer_t {
   char *bytes;
-  uint32_t capacity;
-  uint32_t start;
-  uint32_t size;
+    //uint32_t capacity;
+    //uint32_t start;
+    //uint32_t size;
+    uint64_t capacity;
+    uint64_t start;
+    uint64_t size;
 };
 
 PN_STRUCT_CLASSDEF(pn_buffer)
